pyproject.toml
[project]
name = "gh-to-gitea"
version = "0.1.0"
description = "A CLI tool to mirror GitHub repositories to Gitea."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.5.0",
    "requests>=2.32.3",
    "typer>=0.15.1",
]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[project.scripts]
gh-to-gitea = "gh_to_gitea.main:app"

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]
